+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the "seqcode-preview" extension will be documented in thi
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.0.3 - TBD
+
+- Preview window icon loads.
+- Link paths get workspace folder.
+
+## 0.0.2 - 2025-06-24
+
+- Added dependencies to package.
+- Fixed key bindings.
+
 ## 0.0.1 - 2025-06-23
 
 - Initial release.